moved 'generatePort' from new.js to ajax/helpers.js
